problem: image upload page is confusing
solution: hide file chooser buttons until user activates them for each currency
